Move Old Input bind models
