+++ DCS-SR-Client/Input/Old/InputDevice.cs
@@ -1,7 +1,6 @@
 ﻿using System;
-using Ciribob.DCS.SimpleRadio.Standalone.Client.Settings;
 
-namespace Ciribob.DCS.SimpleRadio.Standalone.Client.Settings
+namespace Ciribob.DCS.SimpleRadio.Standalone.Client.Settings.Old
 {
     public class InputDevice
     {
